Cut the training loader's host memory without changing a batch
Training counts are stored as float32 (bigWig values are float32, so this is exact; the log-count target is still
summed in float64), random crops gather rows in chunks instead of building a full int64 index array, and the
generator keeps one copy of the epoch's examples and drops the previous epoch's before building the next. The
random draws are unchanged: batches are identical to before.

On K562 fold 0 (186k training peaks, 371k nonpeaks) the old loader held 20.3 GiB after loading and peaked at 26.2 GiB
during the per-epoch reshuffle, too close to a 32 GB host once the GPU runtime is added.

+++ chrombpnet/training/utils/data_utils.py
@@ -48,7 +48,8 @@ def get_coords(peaks_df, peaks_bool):
 def get_seq_cts_coords(peaks_df, genome, bw, input_width, output_width, peaks_bool):
 
     seq = get_seq(peaks_df, genome, input_width)
-    cts = get_cts(peaks_df, bw, output_width)
+    # bigWig values are float32: storing them as float32 (not float64) is exact and halves the counts' memory
+    cts = get_cts(peaks_df, bw, output_width).astype(np.float32)
     coords = get_coords(peaks_df, peaks_bool)
     return seq, cts, coords
